+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [11.3.1]
 - [DatePicker] Made sure android only sets date when people tap the ok (positive) button.
 - [TimePicker] Made sure android only sets time when people tap the ok (positive) button.
+- [TimePicker] Made sure android timepicker uses 12/24H format based on the users locale.
 - [HorizontaInlineDatePicker] Made sure it doesnt crash when people select a date from the date picker service.
 - [HorizontaInlineDatePicker] Made sure people can not set a date outside of the upper and lower ranges of the horizontal inline date picker.
 - [DatePickerService] Made sure people have to manually close the date picker to set the date.